Make DsQueryIterator a nested type in DatastoreWrapper and add some Javadoc.
