--- datastore/src/main/java/io/spine/server/storage/datastore/DsQueryIterator.java
@@ -1,88 +1,0 @@
-/*
- * Copyright 2017, TeamDev Ltd. All rights reserved.
- *
- * Redistribution and use in source and/or binary forms, with or without
- * modification, must retain the above copyright notice and the following
- * disclaimer.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package io.spine.server.storage.datastore;
-
-import com.google.cloud.datastore.Cursor;
-import com.google.cloud.datastore.DatastoreReaderWriter;
-import com.google.cloud.datastore.Entity;
-import com.google.cloud.datastore.Query;
-import com.google.cloud.datastore.QueryResults;
-import com.google.cloud.datastore.StructuredQuery;
-
-import java.util.Iterator;
-import java.util.NoSuchElementException;
-
-/**
- * @author Dmytro Dashenkov
- */
-class DsQueryIterator implements Iterator<Entity> {
-
-    private final StructuredQuery<Entity> query;
-    private final DatastoreReaderWriter datastore;
-
-    private QueryResults<Entity> currentPage;
-    private boolean terminated;
-
-    DsQueryIterator(StructuredQuery<Entity> query, DatastoreReaderWriter datastore) {
-        this.query = query;
-        this.datastore = datastore;
-        this.currentPage = datastore.run(query);
-    }
-
-    @Override
-    public boolean hasNext() {
-        if (terminated) {
-            return false;
-        }
-        if (currentPage.hasNext()) {
-            return true;
-        }
-        currentPage = computeNextPage();
-        if (!currentPage.hasNext()) {
-            terminated = true;
-            return false;
-        }
-        return true;
-    }
-
-    @Override
-    public Entity next() {
-        if (!hasNext()) {
-            throw new NoSuchElementException("The query results Iterator is empty.");
-        }
-        final Entity result = currentPage.next();
-        return result;
-    }
-
-    @Override
-    public void remove() {
-        throw new UnsupportedOperationException("Iterator.remove()");
-    }
-
-    private QueryResults<Entity> computeNextPage() {
-        final Cursor cursorAfter = currentPage.getCursorAfter();
-        final Query<Entity> queryForMoreResults = query.toBuilder()
-                                                       .setStartCursor(cursorAfter)
-                                                       .build();
-        final QueryResults<Entity> nextPage = datastore.run(queryForMoreResults);
-        return nextPage;
-    }
-}
